[UPD] Refactoring. Migrate repo, presenter and view interfaces in one FilmActivityContract interface
