+++ app/src/main/java/com/vfalin/sibext/repositories/FilmsActivityRepository.kt
@@ -1,8 +1,39 @@
 package com.vfalin.sibext.repositories
 
 import com.vfalin.sibext.network.pojo.FilmsResponse
+import com.vfalin.sibext.network.service.FilmsService
+import com.vfalin.sibext.ui.activities.FilmsActivityContract
+import io.reactivex.Observer
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.disposables.Disposable
+import io.reactivex.schedulers.Schedulers
+import timber.log.Timber
 
-interface FilmsActivityRepository {
+class FilmsActivityRepository(private val filmsService: FilmsService) :
+    FilmsActivityContract.Model {
 
-    fun getFilms(onSuccess: (FilmsResponse) -> Unit, onError: (Throwable) -> Unit)
+    override fun getFilms(onSuccess: (FilmsResponse) -> Unit, onError: (Throwable) -> Unit) {
+        filmsService.getFilms()
+            .subscribeOn(Schedulers.io())
+            .observeOn(AndroidSchedulers.mainThread())
+            .subscribe(object : Observer<FilmsResponse> {
+                override fun onSubscribe(d: Disposable) {
+                    Timber.i("FilmsActivityRepositoryImpl getFilms Subscription")
+                }
+
+                override fun onNext(response: FilmsResponse) {
+                    onSuccess.invoke(response)
+                    Timber.i("FilmsActivityRepositoryImpl getFilms onNext")
+                }
+
+                override fun onError(throwable: Throwable) {
+                    onError.invoke(throwable)
+                    Timber.i("FilmsActivityRepositoryImpl getFilms onError")
+                }
+
+                override fun onComplete() {
+                    Timber.i("FilmsActivityRepositoryImpl getFilms onComplete")
+                }
+            })
+    }
 }
