+++ app/src/main/java/com/vfalin/sibext/ui/activities/FilmsActivityContract.kt
@@ -1,6 +1,23 @@
 package com.vfalin.sibext.ui.activities
 
+import com.vfalin.sibext.models.FilmsResponseUI
+import com.vfalin.sibext.network.pojo.FilmsResponse
+
 interface FilmsActivityContract {
 
-    fun updateFilms()
+    interface View {
+        fun updateFilms()
+    }
+
+    interface Presenter {
+        fun loadFilms()
+
+        fun getFilms(): FilmsResponseUI
+
+        fun onDestroy()
+    }
+
+    interface Model {
+        fun getFilms(onSuccess: (FilmsResponse) -> Unit, onError: (Throwable) -> Unit)
+    }
 }
